pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "enterprise_mcp_server"
version = "1.3.0"
description = "Enterprise MCP Server with Streamable HTTP transport"
authors = [
    { name = "georgiedekker", email = "georgiedekker@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastmcp>=2.13.0,<3.0.0",
    "fastapi[all]>=0.116.0,<1.0.0",
    "restrictedpython>=8.0,<9.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
    "slowapi>=0.1.9,<0.2.0",
    "sse-starlette>=2.3.4,<3.0.0",
    "pytest>=8.3.5,<9.0.0",
    "jsonify>=0.5,<0.6",
    "httpx>=0.28.1,<0.29.0",
    "pyyaml>=6.0.2,<7.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "asyncpg>=0.31.0,<1.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "pyjwt>=2.10.0,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "redis>=7.0.0,<8.0.0",
    "python-jose[cryptography]>=3.4.0,<4.0.0",
    "python-keycloak>=4.0.0,<6.0.0",
    "mcp>=1.9.0,<2.0.0",
    "claude-code-sdk>=0.0.10,<1.0.0",
    "aiohttp>=3.11.0,<4.0.0",
    "websockets>=15.0.1,<16.0.0",
    "uuid-v7>=1.0.0,<2.0.0",
]
